# feat: reuse HTTP / TCP connection (per cursor)