Addresses https://github.com/python-pinot-dbapi/pinot-dbapi/issues/11

**Before**
``` python
>>> call()
DEBUG:urllib3.connectionpool:Starting new HTTP connection (1): localhost:8099
DEBUG:urllib3.connectionpool:http://localhost:8099 "POST /query/sql HTTP/1.1" 200 696
DEBUG:pinotdb.db:Got the rows as a type <class 'list'> of size 1
DEBUG:pinotdb.db:[['2020-03-21', 6974078886756768268, 'user', 'KPI_10', 35.740856]]

>>> call()
DEBUG:urllib3.connectionpool:Starting new HTTP connection (1): localhost:8099
DEBUG:urllib3.connectionpool:http://localhost:8099 "POST /query/sql HTTP/1.1" 200 696
DEBUG:pinotdb.db:Got the rows as a type <class 'list'> of size 1
DEBUG:pinotdb.db:[['2020-03-21', 6974078886756768268, 'user', 'KPI_10', 35.740856]]

>>> call()
DEBUG:urllib3.connectionpool:Starting new HTTP connection (1): localhost:8099
DEBUG:urllib3.connectionpool:http://localhost:8099 "POST /query/sql HTTP/1.1" 200 696
DEBUG:pinotdb.db:Got the rows as a type <class 'list'> of size 1
DEBUG:pinotdb.db:[['2020-03-21', 6974078886756768268, 'user', 'KPI_10', 35.740856]]
```

**After**
``` python 
>>> call()
DEBUG:urllib3.connectionpool:Starting new HTTP connection (1): localhost:8099    <== Only once
DEBUG:urllib3.connectionpool:http://localhost:8099 "POST /query/sql HTTP/1.1" 200 696
DEBUG:pinotdb.db:Got the rows as a type <class 'list'> of size 1
DEBUG:pinotdb.db:[['2020-03-21', 6974078886756768268, 'user', 'KPI_10', 35.740856]]

>>> call()
DEBUG:urllib3.connectionpool:http://localhost:8099 "POST /query/sql HTTP/1.1" 200 696
DEBUG:pinotdb.db:Got the rows as a type <class 'list'> of size 1
DEBUG:pinotdb.db:[['2020-03-21', 6974078886756768268, 'user', 'KPI_10', 35.740856]]

>>> call()
DEBUG:urllib3.connectionpool:http://localhost:8099 "POST /query/sql HTTP/1.1" 200 696
DEBUG:pinotdb.db:Got the rows as a type <class 'list'> of size 1
DEBUG:pinotdb.db:[['2020-03-21', 6974078886756768268, 'user', 'KPI_10', 35.740856]]
```

**Performance comparison**
Each test was performed 10 times (100 queries) and took the median

**Before**
``` python 
>>> benchmark()
[0] Performing 100 queries to Pinot...
[0] TOOK: 3.092 s
Sleeping...

[1] Performing 100 queries to Pinot...
[1] TOOK: 2.694 s
Sleeping...

[2] Performing 100 queries to Pinot...
[2] TOOK: 2.973 s
Sleeping...

[3] Performing 100 queries to Pinot...
[3] TOOK: 3.074 s
Sleeping...

[4] Performing 100 queries to Pinot...
[4] TOOK: 2.931 s
Sleeping...

[5] Performing 100 queries to Pinot...
[5] TOOK: 2.931 s
Sleeping...

[6] Performing 100 queries to Pinot...
[6] TOOK: 3.173 s
Sleeping...

[7] Performing 100 queries to Pinot...
[7] TOOK: 3.072 s
Sleeping...

[8] Performing 100 queries to Pinot...
[8] TOOK: 3.318 s
Sleeping...

[9] Performing 100 queries to Pinot...
[9] TOOK: 3.217 s
Sleeping...

Median: 3.069 s
```
(Another run: Median = 3.226 s)

**After**
``` python
>>> benchmark()
[0] Performing 100 queries to Pinot...
[0] TOOK: 6.090 s
Sleeping...

[1] Performing 100 queries to Pinot...
[1] TOOK: 2.333 s
Sleeping...

[2] Performing 100 queries to Pinot...
[2] TOOK: 2.257 s
Sleeping...

[3] Performing 100 queries to Pinot...
[3] TOOK: 2.260 s
Sleeping...

[4] Performing 100 queries to Pinot...
[4] TOOK: 2.315 s
Sleeping...

[5] Performing 100 queries to Pinot...
[5] TOOK: 2.216 s
Sleeping...

[6] Performing 100 queries to Pinot...
[6] TOOK: 7.069 s
Sleeping...

[7] Performing 100 queries to Pinot...
[7] TOOK: 2.371 s
Sleeping...

[8] Performing 100 queries to Pinot...
[8] TOOK: 2.309 s
Sleeping...

[9] Performing 100 queries to Pinot...
[9] TOOK: 2.381 s
Sleeping...

Median: 2.328 s
```
(Another run: Median = 2.279 s)

The performance difference is not as significant as I expected (and experienced), so maybe there's a flaw in my measurement strategy.

However, the gain is still significant.

**Script used**
``` python 
import logging
import time
import timeit
import statistics

import pinotdb

logging.basicConfig()
logging.getLogger().setLevel(logging.DEBUG)

client = pinotdb.connect("localhost")
cursor = client.cursor()

def call():
    cursor.execute("SELECT * FROM kpis LIMIT 1").fetchall()

def benchmark(times=10, queries=100):
    logging.getLogger().setLevel(logging.INFO)
    timings = []

    for n in range(times):
        print(f"[{n}] Performing {queries} queries to Pinot...")

        time_taken = timeit.timeit(call, number=queries)
        print(f"[{n}] TOOK: {time_taken:.3f} s")
        timings.append(time_taken)

        print("Sleeping...\n")
        time.sleep(3)

    print(f"Median: {statistics.median_grouped(timings, interval=0.01):.3f} s")
```