@xiangfu0 2021-07-21
> @xiangfu0 What's the release process ?
> So that the updated version is available through PiPy :)

Just released 0.3.4